Loosens the min and max units per project for demoing
If the requirements for finding a fitting selection of projects are 'too tight', the algorithm runs out of iterations before a matching set is found.
This change in the sample data exemplifies that the algorithm does find a selection in reasonable time if the min/max units per project are not too strict.